Package hardening: pyproject packaging, correctness tests, CI, matplotlib 3.10 compatibility
- pyproject.toml with proper metadata (setup.py reduced to a shim; fixes
  the 'Your Name' author placeholder)
- sigmacam/__init__.py now exports the public API and __version__
- boundaries.py: replaced cs.collections (removed in matplotlib 3.10)
  with allsegs, and filter empty contour segments
- utils.py: dtype-safe back-projection in get_mgrid (float32 models no
  longer collide with float64 meshgrids)
- tests/test_sigmacam.py: 10 tests including an analytic correctness
  check - extracted boundary points of a linear-sigmoid model evaluate
  to p=0.5 within 0.02
- CI workflow (CPU torch) runs the suite on every push
- removed placeholder stub files

+++ sigmacam/utils.py
@@ -43,9 +43,10 @@ def get_mgrid(domain, grid_size=100):
     ys = np.linspace(y_min, y_max, grid_size)
     X, Y = np.meshgrid(xs, ys)
     pts2d = np.stack([X.ravel(), Y.ravel()], axis=-1)
-    # back-project
+    # back-project (cast to the domain's dtype so float32 models and
+    # float64 meshgrids can't collide in downstream matmuls)
     P_pinv = torch.linalg.pinv(T)
-    pts_nd = (torch.from_numpy(pts2d) @ P_pinv.T) + x0
+    pts_nd = (torch.from_numpy(pts2d).to(domain.dtype) @ P_pinv.T) + x0
     # mask outside polygon
     mask = Path(domain_2d).contains_points(pts2d)
     grid_points = pts_nd[mask]
